+++ package.json
@@ -13,11 +13,13 @@
     "yargs": "^13.2.4"
   },
   "devDependencies": {
+    "docker-compose": "^0.17.5",
     "ganache-core": "^2.5.5",
     "jest": "^24.8.0",
     "jest-cli": "^24.8.0",
     "jest-junit": "^6.4.0",
     "node-docker-api": "^1.1.22",
+    "node-fetch": "^2.6.0",
     "tar-fs": "^2.0.0"
   },
   "scripts": {
